Reingest AIP: recompress based on reingested compression
refs #6498

Update reingest to parse the compression algorithm out of the reingested
pointer file and use Package.compress_package.

To handle a different compression algorithm resulting in a different
package name (eg .7z vs .tar.bz2) or the AIP being stored in a different
location, do not update the current_path or current_location in the
PackageResource.  Reingest endpoint will reset the current location & path
to match the DB, and pass in the reingest location & path, to be updated
once the package has been successfully reingested.

Squash with 'Stub out finish_reingest AIP'?